src/app/liste/liste.component.ts
import { Component, OnInit, ViewChild } from '@angular/core';
import { MatDialog } from '@angular/material/dialog';
import { EmpAddEditComponent } from '../emp-add-edit/emp-add-edit.component';
import { EmployeeService } from '../services/employee.service';
import { MatPaginator } from '@angular/material/paginator';
import { MatSort } from '@angular/material/sort';
import { MatTableDataSource } from '@angular/material/table';
import { CoreService } from '../core/core.service';
import { ChambreService } from '../chambre.service';


@Component({
  selector: 'app-liste',
  templateUrl: 
  './liste.component.html',

  
  styleUrls: ['./liste.component.scss']
})
export class ListeComponent implements OnInit {
  chambres = [
    { type: 'Single' ,prix:100},
    { type: 'Double' ,prix:150},
    // ... autres chambres avec leurs types
  ];

  displayedColumns: string[] = [
    'id',
    'nom',
    'Prenom',
    'Adresse',
    'email',
    'date_debut',
    'date_fin',
    'Numero_tel',
    'Nombre_chambre',
    'Type_chambre',
    'pension',
    'prix',
    'action',
  ];
  dataSource!: MatTableDataSource<any>;

  @ViewChild(MatPaginator) paginator!: MatPaginator;
  @ViewChild(MatSort) sort!: MatSort;

  constructor(
    private _dialog: MatDialog,
    private _empService: EmployeeService,
    private _coreService: CoreService
  ) {}

  ngOnInit(): void {
    this.getEmployeeList();
  }

  openAddEditEmpForm() {
    const dialogRef = this._dialog.open(EmpAddEditComponent);
    dialogRef.afterClosed().subscribe({
      next: (val) => {
        if (val) {
          this.getEmployeeList();
        }
      },
    });
  }

  getEmployeeList() {
    this._empService.getEmployeeList().subscribe({
      next: (res) => {
        this.dataSource = new MatTableDataSource(res);
        this.dataSource.sort = this.sort;
        this.dataSource.paginator = this.paginator;
      },
      error: console.log,
    });
  }

  

  deleteEmployee(id: number) {
    this._empService.deleteEmployee(id).subscribe({
      next: (res) => {
        this._coreService.openSnackBar('Client supprimé', 'done');
        this.getEmployeeList();
      },
      error: console.log,
    });
  }

  openEditForm(data: any) {
    const dialogRef = this._dialog.open(EmpAddEditComponent, {
      data,
    });

    dialogRef.afterClosed().subscribe({
      next: (val) => {
        if (val) {
          this.getEmployeeList();
        }
      },
    });
  }

  getPrixByType(type: string): number | undefined {
    const chambre = this.chambres.find(chambre => chambre.type === type);
    return chambre ? chambre.prix : undefined;
  }
}
